refactor(file-storage): introduce Store persistence facade
FileService no longer opens connections or runs transactions. A new
Store unit-of-work (infra/storage/store.rs) is the sole owner of the
DBProvider + the three repos + connection-lifecycle and both atomic
operations:

  • bind_atomic — the bind CAS + clear_current + set_current in one
    transaction_ref_mapped (moved verbatim from service.rs §bind).
  • patch_metadata_atomic — the touch_meta + per-key upsert/delete in
    one transaction_ref_mapped (moved verbatim from service.rs §update_metadata).

service.rs drops: toolkit_db::{DBProvider, DbError}, the
toolkit_db::secure::DBRunner bounds, all self.db.conn() calls,
both transaction_ref_mapped blocks, the pending_version fn, and
hash::ALGORITHM. It now holds a single `store: Store` field.

DataPlaneService similarly replaces its db+VersionRepo fields with a
Store clone obtained from FileService::store() — it no longer imports
toolkit_db or VersionRepo directly.

gear.rs constructs the Store once from the DBProvider and passes it
into FileService::new(store, backends, issuer, authorizer, cfg).
service_test.rs updated to match the new constructor.

HK: service.rs 2,322,000 → 619,488 (well under the 1,000,000 target).
Behavior preserved. All tests pass (50 lib + 16 migration + 8 E2E).
Cycle-free; zero new violations.
